// src/api/openapi-client/models/JobPostingCreate.ts
export interface JobPostingCreate {
    /**
     * 
     * @type {string}
     * @memberof JobPostingCreate
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof JobPostingCreate
     */
    startupId: string;
    /**
     * 
     * @type {string}
     * @memberof JobPostingCreate
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof JobPostingCreate
     */
    requiredZip?: string | null;
    /**
     * City where the job must be performed.
     * @type {string}
     * @memberof JobPostingCreate
     */
    requiredCity?: string | null;
    /**
     * Country where job must be performed.
     * @type {string}
     * @memberof JobPostingCreate
     */
    requiredCountry?: string | null;
    /**
     * 
     * @type {number}
     * @memberof JobPostingCreate
     */
    approxDurationWeeks?: number | null;
    /**
     * 
     * @type {number}
     * @memberof JobPostingCreate
     */
    approxHoursPerWeek?: number | null;
    /**
     * 
     * @type {number}
     * @memberof JobPostingCreate
     */
    approxDaysPerWeek?: number | null;
    /**
     * 
     * @type {number}
     * @memberof JobPostingCreate
     */
    approxHourlySalaryEUR?: number | null;
    /**
     * 
     * @type {Date}
     * @memberof JobPostingCreate
     */
    desiredStartDate?: Date | null;
    /**
     * 
     * @type {Array<LanguageProficiency>}
     * @memberof JobPostingCreate
     */
    matchingLanguageProficiencies?: Array<LanguageProficiency>;
    /**
     * 
     * @type {Array<string>}
     * @memberof JobPostingCreate
     */
    matchingExpertiseAreas?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof JobPostingCreate
     */
    matchingSkills?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof JobPostingCreate
     */
    matchingPositions?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof JobPostingCreate
     */
    matchingDegrees?: Array<string>;
}